One Hour Game Jam #019 - Rock
Experimenting with physics particles. A bit of a failure in making a
game but I learned a bit more about how the physics_particle system
behaves. Lessons learned: it doesn't really have a concept of lifetime
built in and deleting older particles manually seems to cause an issue
with the creation of them. Tried a couple of variations but didn't have
luck getting it to work.

Basic idea was to create a volcano which spewed lava which would 'cool'
to rocks. I wanted to use the `physics_particle_` system for the lava
with a lifetime, rather than a deletion region. I might have made an
obvious mistake but the deletion of the particles seems inconsistent.
Tried a few different approaches but kept ending up in the same point.
Out of time, for now, so this will serve as a lesson learned in what not
to do.