One Hour Game Jam #020 - Button
Didn't feel massively inspired by the theme so figured I'd do a game
where you have to go through a maze, and hit a button at the end. First
step, maze generation. There went my time. Figured out the bugs
(variable handling and scope). Added a button to re-trigger maze
generation. Based on a cellular automata maze so not actually a real
maze generation. You can click in the grid to change the cells so a
little game emerges - pick a start and end point and try to place new
cells to re-generate a solvable maze ;)