One Hour Game Jam #021 - Dynamite
Destructible terrain and 'Tanks', or 'Worms', style gameplay by choosing
horizontal and vertical power. Worked out quite well, all things
considered. Didn't have time to put an enemy, or target, in there, but
a good proof of concept and didn't have to put any optimisations in just
for this test of a 4096x768 level.